requirements.txt
evaluate==0.4.0
numpy==1.24.2
PyYAML==6.0
PyYAML==6.0
torch==2.0.0
